NovaAI_WIP/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "nova-ai-wip"
version = "0.1.0"
description = "Experimental AI playground for image search prototypes."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "open_clip_torch",
]

[project.scripts]
nova-ai-wip = "nova_ai_wip.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
